feat(#44): setup config
