+++ constants.py
@@ -24,9 +24,10 @@ class Path:
     MENU_MUSIC_PATH = resource_path(os.path.join(SOUND_PATH, 'menu.wav'))
 
 
-# FONT VARS
-edit_undo_font = resource_path(os.path.join(Path.FONT_PATH, 'edit_undo.ttf'))
-neue_font = resource_path(os.path.join(Path.FONT_PATH, 'neue.ttf'))
+class Font:
+    # FONT VARS
+    edit_undo_font = resource_path(os.path.join(Path.FONT_PATH, 'edit_undo.ttf'))
+    neue_font = resource_path(os.path.join(Path.FONT_PATH, 'neue.ttf'))
 
 
 class Image:
@@ -95,32 +96,3 @@ soundList.append(PLAYER_LASER_SOUND)
 soundList.append(ENEMY_LASER_SOUND)
 soundList.append(EXPLODE_SOUND)
 soundList.append(LASER_HIT_SOUND)
-
-
-class Config:
-    # ROOT VARS
-    TITLE = 'SPACE INVADERS'
-    WIDTH = 750
-    HEIGHT = 750
-
-    # Canvas Dimensions
-    CANVAS = pygame.display.set_mode((WIDTH, HEIGHT))
-
-    screen_rect = CANVAS.get_rect()
-    center_x = screen_rect.centerx
-    starting_x = 0
-    ending_x = WIDTH
-
-    center_y = screen_rect.centery
-    starting_y = 0
-    ending_y = HEIGHT
-
-    # Load Background Image
-    backgroundImage = pygame.image.load(resource_path(os.path.join(
-        Path.GRAPHICS_PATH, 'background-black.png')))
-
-    # Set Background Dimensions
-    BG = pygame.transform.scale(backgroundImage, (WIDTH, HEIGHT))
-
-    FPS = 60
-    framespersec = pygame.time.Clock()
